requirements.txt
Django==1.10
django-tinymce==2.6.0
wsgiref==0.1.2
